+++ changelog.md
@@ -1,4 +1,6 @@
 # v1.1.3
+- Added tooltips to buttons using *[Tooltips](mod:alphalaneous.tooltips)* mod
+- Other tweaks
 
 # v1.1.2
 - Optimized favorites pop-up & mod list UI